Add "Programming Linux Anti-Reversing Techniques" free e-book.
The "Programming Linux Anti-Reversing Techniques" e-book is a free
educational guide designed for junior reverse engineers. It is said to
"teach[…] the reader how to code and analyze well known anti-reversing
techniques for Linux. The book shows how a reverse engineer analyzes a
binary using tools like IDA, Radare2, GDB, readelf, and more. The code
is presented to the reader ready to compile and analyze on their own."